+++ audioswitch/src/test/java/com/twilio/audioswitch/BaseTest.kt
@@ -4,17 +4,24 @@ import android.bluetooth.BluetoothAdapter
 import android.bluetooth.BluetoothClass
 import android.bluetooth.BluetoothDevice
 import android.bluetooth.BluetoothHeadset
+import android.bluetooth.BluetoothProfile
 import android.content.Context
+import android.content.Intent
 import android.media.AudioManager.OnAudioFocusChangeListener
 import com.nhaarman.mockitokotlin2.mock
+import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.whenever
+import com.twilio.audioswitch.AudioDevice.Earpiece
+import com.twilio.audioswitch.AudioDevice.Speakerphone
+import com.twilio.audioswitch.AudioDevice.WiredHeadset
 import com.twilio.audioswitch.android.BuildWrapper
-import com.twilio.audioswitch.android.Logger
 import com.twilio.audioswitch.bluetooth.BluetoothHeadsetManager
 import com.twilio.audioswitch.wired.WiredHeadsetReceiver
+import org.hamcrest.CoreMatchers
+import org.hamcrest.MatcherAssert.assertThat
 
 open class BaseTest {
-    private val bluetoothClass = mock<BluetoothClass> {
+    internal val bluetoothClass = mock<BluetoothClass> {
         whenever(mock.deviceClass).thenReturn(BluetoothClass.Device.AUDIO_VIDEO_HANDSFREE)
     }
     internal val expectedBluetoothDevice = mock<BluetoothDevice> {
@@ -22,7 +29,7 @@ open class BaseTest {
         whenever(mock.bluetoothClass).thenReturn(bluetoothClass)
     }
     internal val context = mock<Context>()
-    internal val logger = mock<Logger>()
+    internal val logger = UnitTestLogger()
     internal val audioManager = setupAudioManagerMock()
     internal val bluetoothAdapter = mock<BluetoothAdapter>()
     internal val audioDeviceChangeListener = mock<AudioDeviceChangeListener>()
@@ -35,6 +42,8 @@ open class BaseTest {
     internal var handler = setupScoHandlerMock()
     internal var systemClockWrapper = setupSystemClockMock()
     internal val headsetProxy = mock<BluetoothHeadset>()
+    internal val preferredDeviceList = listOf(AudioDevice.BluetoothHeadset::class.java, WiredHeadset::class.java,
+            Earpiece::class.java, Speakerphone::class.java)
     internal var headsetManager =
         BluetoothHeadsetManager(
             context, logger, bluetoothAdapter,
@@ -47,6 +56,26 @@ open class BaseTest {
         audioDeviceManager = audioDeviceManager,
         wiredHeadsetReceiver = wiredHeadsetReceiver,
         headsetManager = headsetManager,
-        audioFocusChangeListener = defaultAudioFocusChangeListener
+        audioFocusChangeListener = defaultAudioFocusChangeListener,
+        preferredDeviceList = preferredDeviceList
     )
+
+    internal fun assertBluetoothHeadsetTeardown() {
+        assertThat(headsetManager.headsetListener, CoreMatchers.`is`(CoreMatchers.nullValue()))
+        verify(bluetoothAdapter).closeProfileProxy(BluetoothProfile.HEADSET, headsetProxy)
+        verify(context).unregisterReceiver(headsetManager)
+    }
+
+    internal fun simulateNewBluetoothHeadsetConnection(
+        bluetoothDevice: BluetoothDevice = expectedBluetoothDevice
+    ) {
+        val intent = mock<Intent> {
+            whenever(mock.action).thenReturn(BluetoothHeadset.ACTION_AUDIO_STATE_CHANGED)
+            whenever(mock.getIntExtra(BluetoothHeadset.EXTRA_STATE, BluetoothHeadset.STATE_DISCONNECTED))
+                    .thenReturn(BluetoothHeadset.STATE_CONNECTED)
+            whenever(mock.getParcelableExtra<BluetoothDevice>(BluetoothDevice.EXTRA_DEVICE))
+                    .thenReturn(bluetoothDevice)
+        }
+        headsetManager.onReceive(context, intent)
+    }
 }